// src/app/common/models/asset-type.enum.ts
export enum AssetType {
    Cash_Account = 1,
    Superannuation_Account = 2,
    Retirement_Income_Stream = 3,
    Investment_Fund = 4,
    Direct_Shares = 100000002,
    Property = 100000001,
    Centrelink_Benefit = 100000000,
    Legacy = 7,
    Other = 8,
    Whole_of_Life = 509000000,
    Endowment = 509000001,
    Insurance_Bond = 509000002,
    Fixed_Interest = 509000003,
    
}